Refactor blog components
