+++ components/blog/AuthorBio.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import Link from 'next/link';
 import ExportedImage from 'next-image-export-optimizer';
 import { Markup } from 'react-render-markup';
