+++ next.config.js
@@ -1,4 +1,5 @@
 const withPWA = require('next-pwa')({
+  disable: process.env.NODE_ENV === 'development',
   dest: 'public',
   register: true,
   skipWaiting: true,
